# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dorshi Holiday Resort Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  timestamp?: string;
}

export interface Room {
  id: string;
  name: string;
  type: string;
  capacity: number;
  pricePerNight: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  images?: string[] | null;
}

export interface BookingInput {
  guestName: string;
  phone: string;
  roomId: string;
  checkIn: string;
  checkOut: string;
  guests: number;
}

export interface BookingRecord {
  id: string;
  guestName: string;
  phone: string;
  roomId: string;
  checkIn: string;
  checkOut: string;
  guests: number;
  status: string;
  /** @nullable */
  createdAt?: string | null;
}

export interface BookingResult {
  success: boolean;
  booking?: BookingRecord;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  retry?: boolean | null;
}

export interface CancellationRequest {
  action: string;
  /** @nullable */
  guestName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bookingId?: string | null;
}

export interface CancellationResponse {
  found: boolean;
  booking?: BookingRecord;
  /** @nullable */
  success?: boolean | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  bookingId?: string | null;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetAvailableRoomsParams = {
checkIn: string;
checkOut: string;
};