# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dorshi Holiday Resort Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: getAvailableRooms
      tags: [rooms]
      summary: Get available rooms for date range
      parameters:
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of available rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, name, type, capacity, pricePerNight]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        capacity:
          type: integer
        pricePerNight:
          type: integer
        description:
          type: ["string", "null"]
        images:
          type: ["array", "null"]
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string