# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Dorshi Holiday Resort Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  name: string;
  type: string;
  capacity: number;
  pricePerNight: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  images?: string[] | null;
}